requirements.txt
# ─── Django Core ───────────────────────────────────────────────────────────────
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-filter==24.2
django-taggit==5.0.1

# ─── Authentication ────────────────────────────────────────────────────────────
djangorestframework-simplejwt==5.3.1

# ─── Database ──────────────────────────────────────────────────────────────────
dj-database-url==2.1.0
psycopg2-binary==2.9.9

# ─── Async / Task Queue ────────────────────────────────────────────────────────
celery==5.3.6
redis==5.0.3
django-celery-beat==2.6.0
django-celery-results==2.5.1
channels==4.0.0
channels-redis==4.2.0

# ─── HTTP / Scraping ───────────────────────────────────────────────────────────
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
yt-dlp==2024.3.10

# ─── Image Processing ──────────────────────────────────────────────────────────
Pillow==10.3.0
pdf2image==1.17.0

# ─── Speech-to-Text ────────────────────────────────────────────────────────────
openai-whisper==20231117

# ─── Full-Text Search ──────────────────────────────────────────────────────────
Whoosh==2.7.4

# ─── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
python-magic==0.4.27
gunicorn==21.2.0
whitenoise==6.6.0

# ─── Development / Testing ─────────────────────────────────────────────────────
pytest==8.1.1
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.0
faker==24.4.0
django-extensions==3.2.3
ipython==8.23.0
